% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.alpha}
\alias{as.alpha}
\title{as.alpha}
\usage{
as.alpha(nmf.LassoK.result)
}
\arguments{
\item{nmf.LassoK.result}{discovered signatures into an object as the one generated by the function nmf.LassoK}
}
\value{
A matrix with alpha values
}
\description{
Return the discovered alpha values from an input object as the one generated by the function nmf.LassoK.
}
\examples{
data(nmf_LassoK_example)
res = as.alpha(nmf_LassoK_example)

}
