% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.beta.in.range}
\alias{as.beta.in.range}
\title{as.beta.in.range}
\usage{
as.beta.in.range(lambda.range.result, lambda_value = 0.2)
}
\arguments{
\item{lambda.range.result}{discovered signatures into an object as the one generated by the function evaluate.lambda.range}

\item{lambda_value}{value of LASSO used for the estimation whose configuration has to be returned}
}
\value{
A matrix with beta values
}
\description{
Return the discovered signatures (beta values) for a given configuration from an input object as the one generated by the function evaluate.lambda.range.
}
