% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.beta}
\alias{as.beta}
\title{as.beta}
\usage{
as.beta(nmf.LassoK.result)
}
\arguments{
\item{nmf.LassoK.result}{discovered signatures into an object as the one generated by the function nmf.LassoK}
}
\value{
A matrix with beta values
}
\description{
Return the discovered signatures (beta values) from an input object as the one generated by the function nmf.LassoK.
}
\examples{
data(nmf_LassoK_example)
res = as.beta(nmf_LassoK_example)

}
