% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.utils.R
\name{as.loglik.progression.in.range}
\alias{as.loglik.progression.in.range}
\title{as.loglik.progression.in.range}
\usage{
as.loglik.progression.in.range(lambda.range.result, lambda_value = 0.2)
}
\arguments{
\item{lambda.range.result}{discovered signatures into an object as the one generated by the function evaluate.lambda.range}

\item{lambda_value}{value of LASSO used for the estimation whose configuration has to be returned}
}
\value{
A vector with log-likelihood values for each iteration during the estimation of the signatures
}
\description{
Return the log-likelihood values for a given configuration for each iteration during the estimation of the signatures from an input object as the one 
generated by the function evaluate.lambda.range.
}
\examples{
data(lambda_range_example)
res = as.loglik.progression.in.range(lambda_range_example,lambda_value=0.10)

}
