% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.utils.R
\name{as.loglik.progression}
\alias{as.loglik.progression}
\title{as.loglik.progression}
\usage{
as.loglik.progression(nmf.LassoK.result)
}
\arguments{
\item{nmf.LassoK.result}{discovered signatures into an object as the one generated by the function nmf.LassoK}
}
\value{
A vector with log-likelihood values for each iteration during the estimation of the signatures
}
\description{
Return the log-likelihood values for each iteration during the estimation of the signatures from an input object as the one generated by 
the function nmf.LassoK.
}
\examples{
data(nmf_LassoK_example)
res = as.loglik.progression(nmf_LassoK_example)

}
