% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.mean.squared.error}
\alias{as.mean.squared.error}
\title{as.mean.squared.error}
\usage{
as.mean.squared.error(nmf.LassoCV.result)
}
\arguments{
\item{nmf.LassoCV.result}{results of cross validation into an object as the one generated by the function nmf.LassoCV}
}
\value{
A list with distributions and median values of the mean squared error values computed between the observed counts 
and the predicted ones at each iteration of cross validation
}
\description{
Return the mean squared error between the observed counts and the predicted ones for all the configurations considered during cross 
validation. This function expects an input object as the one generated by the function nmf.LassoCV.
}
