% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.utils.R
\name{as.starting.beta.in.range}
\alias{as.starting.beta.in.range}
\title{as.starting.beta.in.range}
\usage{
as.starting.beta.in.range(lambda.range.result, lambda_value = 0.2)
}
\arguments{
\item{lambda.range.result}{discovered signatures into an object as the one generated by the function nmf.LassoK}

\item{lambda_value}{value of LASSO used for the estimation whose configuration has to be returned}
}
\value{
A matrix with the initial values of beta used for the estimation of the signatures
}
\description{
Return the initial values of beta used for the estimation of the signatures for a given configuration from an input object as the one generated by the 
function evaluate.lambda.range.
}
\examples{
data(lambda_range_example)
res = as.starting.beta.in.range(lambda_range_example,lambda_value=0.10)

}
