% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.starting.beta}
\alias{as.starting.beta}
\title{as.starting.beta}
\usage{
as.starting.beta(nmf.LassoK.result)
}
\arguments{
\item{nmf.LassoK.result}{discovered signatures into an object as the one generated by the function nmf.LassoK}
}
\value{
A matrix with the initial values of beta used for the estimation of the signatures
}
\description{
Return the initial values of beta used for the estimation of the signatures from an input object as the one generated by the function nmf.LassoK.
}
