% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background}
\alias{background}
\title{germline replication error}
\format{
vector of rates
}
\source{
Rahbari, Raheleh, et al. "Timing, rates and spectra of human germline mutation." Nature genetics 48.2 (2016): 126.
}
\usage{
data(background)
}
\value{
vector of rates for the 96 trinucleotides
}
\description{
germline replication error estimated in Rahbari, Raheleh, et al. (2016).
}
