% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import.counts.data}
\alias{import.counts.data}
\title{import.counts.data}
\usage{
import.counts.data(input, bsg, mutation_categories)
}
\arguments{
\item{input}{either a data.frame/data.table object or a file with 5 columns: sample name, chromosome, position, ref, alt.}

\item{bsg}{a BSgenome object for the reference genome. Chromosome names have to match the input table.}

\item{mutation_categories}{array with the 96 mutational categories to be considered. It is provided along with the package 
by data(mutation_categories).}
}
\value{
A count matrix to extract mutatational signatues
}
\description{
Import point mutations data to build the count matrix to extract mutatational signatues.
}
\examples{
data(imported_data)
head(imported_data)

}
