% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imported_data}
\alias{imported_data}
\title{example of imported data from Nik-Zainal, Serena, et al. (2016).}
\format{
results obtained with the function import.counts.data on the data from Nik-Zainal, Serena, et al. (2016)
}
\usage{
data(imported_data)
}
\value{
results obtained with the function import.counts.data on the data from Nik-Zainal, Serena, et al. (2016)
}
\description{
example of imported data from Nik-Zainal, Serena, et al. (2016).
}
