% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutation_categories}
\alias{mutation_categories}
\title{trinucleotides mutation categories}
\format{
matrix of 96 trinucleotides mutation categories
}
\usage{
data(mutation_categories)
}
\value{
matrix of 96 trinucleotides mutation categories
}
\description{
96 trinucleotides mutation categories
}
