% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmf_LassoK_example}
\alias{nmf_LassoK_example}
\title{example of results obtained with the function nmf.LassoK on the counts input from Nik-Zainal, Serena, et al. (2016).}
\format{
results obtained with the function nmf.LassoK on the counts input from Nik-Zainal, Serena, et al. (2016)
}
\usage{
data(nmf_LassoK_example)
}
\value{
results obtained with the function nmf.LassoK on the counts input from Nik-Zainal, Serena, et al. (2016)
}
\description{
example of results obtained with the function nmf.LassoK on the counts input from Nik-Zainal, Serena, et al. (2016).
}
