% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{patients.plot}
\alias{patients.plot}
\title{patients.plot}
\usage{
patients.plot(
  trinucleotides_counts,
  samples = rownames(trinucleotides_counts),
  freq = FALSE,
  xlabels = FALSE
)
}
\arguments{
\item{trinucleotides_counts}{trinucleotides counts matrix.}

\item{samples}{name of the samples. This should match a rownames in trinucleotides_counts.}

\item{freq}{boolean value; shall I display rates instead of counts?}

\item{xlabels}{boolean value; shall I display x labels?}
}
\value{
A ggplot2 object.
}
\description{
Plot trinucleotides counts for a set of given patients.
}
\examples{
data(patients)
patients.plot(trinucleotides_counts=patients,samples=c("PD10010a","PD10011a","PD10014a"))

}
