% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patients}
\alias{patients}
\title{point mutations for 560 breast tumors}
\format{
counts of the point mutations
}
\source{
Nik-Zainal, Serena, et al. "Landscape of somatic mutations in 560 breast cancer whole-genome sequences." Nature 534.7605 (2016): 47.
}
\usage{
data(patients)
}
\value{
counts of point mutations for 560 tumors and 96 trinucleotides
}
\description{
dataset of counts of the point mutations detected in 560 breast tumors published in Nik-Zainal, Serena, et al. (2016).
}
