% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.patient}
\alias{plot.patient}
\title{plot.patient}
\usage{
plot.patient(countMatrix, patientName, xlabels = TRUE, freq = FALSE)
}
\arguments{
\item{countMatrix}{count matrix as the one generated by the import.data function.}

\item{patientName}{name of the patient. This should match a rowname of the countMatrix.}

\item{xlabels}{boolean value; shall I display x labels?}

\item{freq}{boolean value; shall I display rates instead of counts?}
}
\value{
A ggplot2 object
}
\description{
Plot the counts for a patient.
}
