% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot.signatures}
\alias{plot.signatures}
\title{plot.signatures}
\usage{
plot.signatures(beta, useColNames = TRUE, firstBackground = TRUE,
  xlabels = TRUE)
}
\arguments{
\item{beta}{discovered signatures}

\item{useColNames}{boolean value; shall I use the colnames from beta as names for the signatures?}

\item{firstBackground}{boolean value; shall I display the background signature as the first element?}

\item{xlabels}{boolean value; shall I display x labels?
by data(mutation_categories).}
}
\value{
A ggplot2 object
}
\description{
Plot the discovered signatures.
}
