% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{signatures.plot}
\alias{signatures.plot}
\title{signatures.plot}
\usage{
signatures.plot(
  beta,
  useColNames = TRUE,
  mutation_categories = NULL,
  firstBackground = TRUE,
  xlabels = TRUE
)
}
\arguments{
\item{beta}{discovered signatures}

\item{useColNames}{boolean value; shall I use the colnames from beta as names for the signatures?}

\item{mutation_categories}{if useColNames is FALSE, the trinucleotides categories to be considered can be specified. 
An example is provided in the package and can be loaded by data(mutation_categories)}

\item{firstBackground}{boolean value; shall I display the background signature as the first element?}

\item{xlabels}{boolean value; shall I display x labels?}
}
\value{
A ggplot2 object
}
\description{
Plot the discovered signatures.
}
\examples{
data(nmf_LassoK_example)
beta = as.beta(nmf_LassoK_example)
signatures.plot(beta=beta)

}
