% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssm560}
\alias{ssm560}
\title{point mutations for 560 breast tumors in the format compatible with the import function}
\format{counts of the point mutations in the format compatible with the import function}
\source{
Nik-Zainal, Serena, et al. "Landscape of somatic mutations in 560 breast cancer whole-genome sequences." Nature 534.7605 (2016): 47.
}
\usage{
data(ssm560)
}
\value{
counts of point mutations for 560 tumors and 96 trinucleotides in the format compatible with the import function
}
\description{
dataset of counts of the point mutations detected in 560 breast tumors published in Nik-Zainal, Serena, et al. (2016).
}
