% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starting_betas_example}
\alias{starting_betas_example}
\title{example of results obtained with the function starting.betas.estimation on the counts input from Nik-Zainal, Serena, et al. (2016).}
\format{
results obtained with the function starting.betas.estimation on the counts input from Nik-Zainal, Serena, et al. (2016)
}
\usage{
data(starting_betas_example)
}
\value{
results obtained with the function starting.betas.estimation on the counts input from Nik-Zainal, Serena, et al. (2016)
}
\description{
example of results obtained with the function starting.betas.estimation on the counts input from Nik-Zainal, Serena, et al. (2016).
}
